feat: integrate PostgreSQL with pgvector and JPA for document storage
- Add PostgreSQL and pgvector JDBC dependencies for production
- Enable JPA repositories and entity scanning in DocumentServiceApplication
- Refactor DocumentService from in-memory ConcurrentHashMap to DocumentRepository
- Change semantic scoring from int to float for precision with vector embeddings

+++ backend/document-service/src/main/java/com/example/fhir/service/AiMetadataService.java
@@ -77,7 +77,7 @@ public class AiMetadataService {
         return enrichLocally(request);
     }
 
-    public int semanticScore(ClinicalDocument document, String query) {
+    public float semanticScore(ClinicalDocument document, String query) {
         if (query == null || query.isBlank()) {
             return 0;
         }
@@ -110,7 +110,7 @@ public class AiMetadataService {
         return new EnrichedMetadata(title, summary, keywords);
     }
 
-    private int semanticScoreLocally(ClinicalDocument document, String query) {
+    private float semanticScoreLocally(ClinicalDocument document, String query) {
         String normalizedQuery = query.toLowerCase(Locale.ROOT);
         int score = 0;
         score += containsScore(document.getTitle(), normalizedQuery, 8);
@@ -232,6 +232,6 @@ public class AiMetadataService {
     private record SidecarEnrichmentResponse(String title, String summary, List<String> keywords) {
     }
 
-    private record SidecarScoreResponse(int score) {
+    private record SidecarScoreResponse(float score) {
     }
 }
